requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Data Processing
pyarrow>=6.0.0
fastparquet>=0.7.0

# Machine Learning
scikit-learn>=1.0.0
torch>=1.10.0
torchvision>=0.11.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Notebook
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Utilities
tqdm>=4.62.0
pyyaml>=5.4.0

# Optional: Hyperparameter Optimization
optuna>=2.10.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0

# Development
black>=21.0
flake8>=3.9.0
mypy>=0.910
